+++ config/defaults.inc.php
@@ -883,9 +883,9 @@ $config['request_path'] = null;
 // One replacement variable is supported: %n (resolves to the value of $_SERVER['SERVER_NAME']).
 $config['request_url'] = null;
 
-// Allows to define separate server/path for image/js/css files
-// Warning: If the domain is different cross-domain access to some
-// resources need to be allowed
+// Allows to define separate URL/path for image/js/css files. It has to be a full URL
+// or an absolute path (starting with /).
+// Warning: If the domain is different cross-domain access to some resources need to be allowed
 // Sample:
 //    <FilesMatch ".(eot|ttf|woff)">
 //    Header set Access-Control-Allow-Origin "*"