+++ program/js/app.js
@@ -10629,7 +10629,7 @@ function rcube_webmail() {
             path = this.env.assets_path + path;
         }
 
-        if (!path.startsWith('static.php/') && path.indexOf('://') == -1) {
+        if (!path.startsWith('static.php/') && path.indexOf('://') == -1 && path.charAt(0) != '/') {
             path = 'static.php/' + path;
         }